Use u64 for trade message shares
This is necessary because cross trades use u64 for shares, whereas trades use u32. We needed to use the larger when serializing into a single trade type.

+++ src/message.rs
@@ -265,7 +265,7 @@ pub struct TradeMessage {
     kind: char, // P = non-cross, Q = cross, B = broken
     refno: u64,
     side: Side, // The type of non-display order on the book being matched (always "B" effective 07/14/2014)
-    shares: u32,
+    shares: u64,
     ticker: String,
     price: u32,
     matchno: u64,
